fix minor bug

+++ src/Husky.Helpers/Helpers/LatLonHelper.cs
@@ -18,10 +18,10 @@ namespace Husky
 			if ( one.LatLonType != another.LatLonType ) {
 				throw new ArgumentException("These two locations are in different coordinate standards");
 			}
-			var radLat1 = one.Lat * Math.PI / 180.0;
-			var radLat2 = another.Lat * Math.PI / 180.0;
+			var radLat1 = (double)one.Lat * Math.PI / 180.0;
+			var radLat2 = (double)another.Lat * Math.PI / 180.0;
 			var a = radLat1 - radLat2;
-			var b = one.Lon * Math.PI / 180.0 - another.Lon * Math.PI / 180.0;
+			var b = (double)one.Lon * Math.PI / 180.0 - (double)another.Lon * Math.PI / 180.0;
 			var s = 2 * Math.Asin(Math.Sqrt(Math.Pow(Math.Sin(a / 2), 2) + Math.Cos(radLat1) * Math.Cos(radLat2) * Math.Pow(Math.Sin(b / 2), 2)));
 			return (int)Math.Round(s * 6378.137 * 1000);
 		}
@@ -31,11 +31,11 @@ namespace Husky
 				latlon.LatLonType = LatLonType.Bd09;
 				return latlon;
 			}
-			var z = Math.Sqrt(latlon.Lon * latlon.Lon + latlon.Lat * latlon.Lat) + 0.00002 * Math.Sin(latlon.Lat * xPI);
-			var theta = Math.Atan2(latlon.Lat, latlon.Lon) + 0.000003 * Math.Cos(latlon.Lon * xPI);
+			var z = Math.Sqrt((double)(latlon.Lon * latlon.Lon + latlon.Lat * latlon.Lat)) + 0.00002 * Math.Sin((double)latlon.Lat * xPI);
+			var theta = Math.Atan2((double)latlon.Lat, (double)latlon.Lon) + 0.000003 * Math.Cos((double)latlon.Lon * xPI);
 			return new Location {
-				Lat = z * Math.Sin(theta) + 0.006,
-				Lon = z * Math.Cos(theta) + 0.0065,
+				Lat = (decimal)(z * Math.Sin(theta) + 0.006),
+				Lon = (decimal)(z * Math.Cos(theta) + 0.0065),
 				LatLonType = LatLonType.Bd09
 			};
 		}
@@ -45,13 +45,13 @@ namespace Husky
 				latlon.LatLonType = LatLonType.Gcj02;
 				return latlon;
 			}
-			var lat = latlon.Lat - 0.006;
-			var lon = latlon.Lon - 0.0065;
+			var lat = (double)latlon.Lat - 0.006;
+			var lon = (double)latlon.Lon - 0.0065;
 			var z = Math.Sqrt(lon * lon + lat * lat) - 0.00002 * Math.Sin(lat * xPI);
 			var theta = Math.Atan2(lat, lon) - 0.000003 * Math.Cos(lon * xPI);
 			return new Location {
-				Lat = z * Math.Sin(theta),
-				Lon = z * Math.Cos(theta),
+				Lat = (decimal)(z * Math.Sin(theta)),
+				Lon = (decimal)(z * Math.Cos(theta)),
 				LatLonType = LatLonType.Gcj02
 			};
 		}
